requirements.txt
pyodbc
pandas
google-cloud-bigquery
pandas-gbq
python-dotenv